extract megaphone to own UIView subclass; adaptive scaling of vertical padding
